feat: タスク操作ボタンのレスポンシブ修正 #180
