+++ frontend/src/components/models/task/DeleteTasksButton.tsx
@@ -1,24 +1,43 @@
+import { Dispatch, SetStateAction, useState } from "react";
 import { IconButton, Button } from "@mui/material";
+import { GridRowId } from "@mui/x-data-grid";
 import DeleteIcon from "@mui/icons-material/Delete";
 import { AlertDialog } from "../../ui/AlertDialog";
+import { useDeleteTasks } from "../../../hooks/useDeleteTasks";
 
 type Props = {
-  handleOpen: () => void;
-  handleClose: () => void;
-  disabled: boolean;
-  open: boolean;
-  onClick: () => void;
+  selectionModel: GridRowId[];
+  isFinished: boolean;
+  flag: boolean;
+  setFlag: Dispatch<SetStateAction<boolean>>;
 };
 
 export const DeleteTasksButton = (props: Props) => {
-  const { handleOpen, handleClose, disabled, open, onClick } = props;
+  const { selectionModel, isFinished, flag, setFlag } = props;
+  const [open, setOpen] = useState(false);
+
+  const handleOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  const handleDeleteTasks = useDeleteTasks({
+    selectionModel,
+    isFinished,
+    flag,
+    setFlag,
+    handleClose,
+  });
 
   return (
     <div>
       <IconButton
         color="error"
         onClick={handleOpen}
-        disabled={disabled}
+        disabled={selectionModel?.length === 0}
         sx={{ display: { xs: "flex", md: "none" } }}
       >
         <DeleteIcon />
@@ -27,7 +46,7 @@ export const DeleteTasksButton = (props: Props) => {
         color="error"
         variant="outlined"
         onClick={handleOpen}
-        disabled={disabled}
+        disabled={selectionModel?.length === 0}
         startIcon={<DeleteIcon />}
         sx={{ display: { xs: "none", md: "flex" } }}
       >
@@ -37,7 +56,7 @@ export const DeleteTasksButton = (props: Props) => {
         open={open}
         handleClose={handleClose}
         objectName="タスク"
-        onClick={onClick}
+        onClick={handleDeleteTasks}
       />
     </div>
   );
