+++ frontend/src/hooks/usePatchTasks.ts
@@ -12,15 +12,14 @@ type Props = {
   isFinished: boolean;
   flag: boolean;
   setFlag: Dispatch<SetStateAction<boolean>>;
-  handleClose: () => void;
 };
 
-export const useHandleMultiTasks = (props: Props) => {
-  const { selectionModel, isFinished, flag, setFlag, handleClose } = props;
+export const usePatchTasks = (props: Props) => {
+  const { selectionModel, isFinished, flag, setFlag } = props;
   const { teamReloadFlag, setTeamReloadFlag } = useContext(AuthContext);
   const { handleSetSnackbar } = useContext(SnackbarContext);
 
-  const handleMultiIsDoneUpdate = useCallback(() => {
+  const handleUpdateIsDoneTasks = useCallback(() => {
     // eslint-disable-next-line array-callback-return
     selectionModel?.map((id) => {
       const options: AxiosRequestConfig = {
@@ -66,42 +65,5 @@ export const useHandleMultiTasks = (props: Props) => {
     });
   }, [selectionModel, isFinished, flag, teamReloadFlag]);
 
-  const handleMultiTasksDelete = useCallback(() => {
-    // eslint-disable-next-line array-callback-return
-    selectionModel?.map((id) => {
-      const options: AxiosRequestConfig = {
-        url: `/tasks/${id}`,
-        method: "DELETE",
-        headers: {
-          "access-token": Cookies.get("_access_token") || "",
-          client: Cookies.get("_client") || "",
-          uid: Cookies.get("_uid") || "",
-        },
-      };
-
-      baseAxios(options)
-        .then((res: AxiosResponse) => {
-          console.log(res);
-          setFlag(!flag);
-          setTeamReloadFlag(!teamReloadFlag);
-          handleSetSnackbar({
-            open: true,
-            type: "success",
-            message: "タスクを削除しました",
-          });
-          handleClose();
-        })
-        .catch((err) => {
-          console.log(err);
-          handleSetSnackbar({
-            open: true,
-            type: "error",
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment
-            message: err.response.data.errors,
-          });
-        });
-    });
-  }, [selectionModel, isFinished, flag, teamReloadFlag]);
-
-  return { handleMultiIsDoneUpdate, handleMultiTasksDelete };
+  return handleUpdateIsDoneTasks;
 };
